refactor: #2 enums 패키지 생성
