+++ src/main/java/com/study/simpleboard/domain/enums/ReactionType.java
@@ -1,4 +1,4 @@
-package com.study.simpleboard.domain;
+package com.study.simpleboard.domain.enums;
 
 public enum ReactionType {
     LIKE,
