+++ src/main/java/com/study/simpleboard/domain/enums/TargetType.java
@@ -1,4 +1,4 @@
-package com.study.simpleboard.domain;
+package com.study.simpleboard.domain.enums;
 
 public enum TargetType {
     POST,
